src/requirements.txt
fastapi[standard]==0.115.2
pydantic-settings==2.6.0
aiofiles==24.1.0
langchain==0.3.9
PyMuPDF==1.24.12
motor==3.6.0
langchain-core==0.3.21
langchain-community==0.3.3
openai==1.60.2
cohere==5.13.11
qdrant-client==1.13.2
google-genai==1.0.0
SQLAlchemy==2.0.38
asyncpg==0.30.0
alembic==1.14.1
psycopg2==2.9.10
psycopg==3.2.4
pgvector==0.4.1
nltk==3.9.1

